+++ .upgrade.yml
@@ -1,6 +1,6 @@
 mappings:
-  DocumentConverterDecorator: SilverStripe\DocumentConverter\DocumentConverter
-  DocumentImportField: SilverStripe\DocumentConverter\DocumentImportField
-  DocumentImportIFrameField_Importer: SilverStripe\DocumentConverter\DocumentImportIFrameFieldImporter
-  DocumentImportInnerField: SilverStripe\DocumentConverter\DocumentImportInnerField
+  DocumentConverterDecorator: SilverStripe\DocumentConverter\ConvertsDocumentsExtension
+  DocumentImportField: SilverStripe\DocumentConverter\DocumentConversionField
+  DocumentImportIFrameField_Importer: SilverStripe\DocumentConverter\DocumentConverter
+  DocumentImportInnerField: SilverStripe\DocumentConverter\DocumentImporterField
   DocumentConverterTest: SilverStripe\DocumentConverter\Tests\DocumentConverterTest
