+++ composer.json
@@ -18,6 +18,9 @@
         }
     },
 	"extra": {
+        "expose": [
+            "javascript"
+        ],
 		"branch-alias": {
 			"dev-master": "2.x-dev"
 		}
