+++ lang/de.yml
@@ -1,7 +1,7 @@
 de:
-  SilverStripe\DocumentConverter\DocumentConverter:
+  SilverStripe\DocumentConverter\ConvertsDocumentsExtension:
     ImportTab: Importieren
-  DocumentImportField:
+  SilverStripe\DocumentConverter\DocumentConversionField:
     ChooseFolder: 'Wählen Sie einen Ordner zum Speichern dieser Datei aus'
     EachH1: 'für jede Überschrift 1'
     EachH2: 'für jede Überschrift 2'
