+++ lang/mi.yml
@@ -2,7 +2,7 @@ mi:
   DocumentConverter:
     PROCESSFAILED: 'Kāore i taea te tukatuka te tuhinga, whakatūturu pūtia kua tukuna atu he hōputu .doc, .docx rānei.'
     SERVERUNREACHABLE: 'Kāore i taea te whakapā ki te tūmau huri tuhinga. Me ngana anō ā muri ake, ka whakapā rānei ki tō kaiwhakahaere pūnaha.'
-  DocumentCSilverStripe\DocumentConverter\DocumentConverteronverterDecorator:
+  DocumentCSilverStripe\DocumentConverter\ConvertsDocumentsExtension:
     ImportTab: Kawemai
   FileIFrameField:
     ATTACHONCESAVED2: 'Ka taea te āpiti kōnae ina oti te tiaki tuatahi o te pūkete.'
